QmlJSEditor: Fix running QML/JS checks
For non-CMake based projects with enabled qmlls the checks
just do nothing as the qmllint target does not get created
elsewhere and an early return prevents the fallback in such
case.
Limit the usage of qmllint to cmake based projects to be able
to use the fallback for other build systems.
Regression introduced with 1b57e95c14d78119bbf8358bb52dfc1be0cde140.

Change-Id: Ie22765aecea2816560488d21b2d355ba96ff3589
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>