Android: Get rid of packageReloadBegin() signal
Since the packageReloadBegin() signal is always followed by the
packageReloadFinished() signal, emitted synchronously to the former one,
there is no need to have 2 separate signals. Particularly,
showing the progress on packageReloadBegin() and hiding it on
packageReloadFinished() is no-op, since the gui won't have a chance to
paint it (because of synchronous emission of both signals).

Thus, remove the packageReloadBegin() signal and remove the progress
indicator.

Rename packageReloadFinished() to packagesReloaded().

Inline clearContainers().

Change-Id: I0ad65dd4d915a1d164c8d398a84a568c57545a6f
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>