Utils: fix conflict of withSuffix(int) and withSuffix(qsizetype)
qsizetype is equivalent to ssize_t, meaning signed 32bit on 32bit
architectures, and thus this overload conflicts with
withSuffix(qsizetype) version on 32bit architectures.

Also, as qsizetype is a 64bit type on 64bit architectures, the fact that
withSuffix(qsizetype) calls withSuffix(int) by downcasting the
parameter to int means that there may be a truncation/loss in case the
value is large enough.

Hence, to fix both the issues:
- switch the two implementations, so now withSuffix(int) calls
  withSuffix(qsizetype)
- disable disable withSuffix(int) on 32bit architectures, as its job is
  already done by withSuffix(qsizetype)

Change-Id: Icb82d2ba58362cc10f18cf2c41ad409c39b60863
Reviewed-by: hjk <hjk@qt.io>