VcsBase: Fix enabling git actions after rebase
Delay the state changed signal for proper evaluation
(a file check for `rebase-apply` is run when the
signal is emitted)

This fixes that Git "Rebasing" stayed in the title
bar when rebasing was long finished and therefore
the rebase actions stayed disabled.

TODO: The [REBASING] message in the projects view is
not fixed, it disappears when clicking the title bar.

Fixes: QTCREATORBUG-31815
Change-Id: Ic83949dfdaf26df3af8149457e3f282ff1bcefe5
Reviewed-by: Orgad Shaneh <orgads@gmail.com>