qmljscheck: fix M129 warning on components with same name
Look in the imports up whether a component with the same name as the
current component was imported. In this case, this component shadows the
current component and can be used in the file without recursively
instantiating anything.

Therefore, count the number of components with the same name as the
currently edited qml file that was imported in the current file and warn
if no other component with the same name was imported.

Fixes: QTCREATORBUG-31313
Change-Id: Id1e748dc0dcdbb514d9882cf79dfba959abacd15
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>