QmlDesigner: Add generated components when exporting a bundle
When exporting a bundle of a component that has a dependency on another
component loaded via a lib import from the Generated folder, those
dependency components get properly added to the exported bundle. Also
the generated imports are removed from the main component as they are
not valid anymore.

Task-number: QDS-13746
Change-Id: Ic3370fc62baa6b2a7cbbef432b1c2a3853170250
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>