PE: Fix Kit enablement
The TargetSetupWidget::toggleEnabled(), called first, sets the enablement
for TargetSetupWidget's subcontrols, but not for TargetSetupWidget
itself.

In the scenario from the bugreport the toggleEnabled() was called
with false, but the enablement of TargetSetupWidget was still true.
Later, when selectAtLeastOneEnabledKit() is called, the internals
take TargetSetupWidget's enablement into account, not the state
set with previous call to toggleEnabled().

Rename toggleEnabled() into setValid(), store the flag internally,
and add a getter isValid(). Reuse this getter in
selectAtLeastOneEnabledKit().

Fixes: QTCREATORBUG-31879
Change-Id: I7b9be718cc6732afab6eb26387b2a88701a7fd67
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: hjk <hjk@qt.io>