Android: fix wording in the Android templates wizard
All strings here still refer only to qmake projects,
this is no longer accurate of course, so change the
wording to "build target" instead of ".pro file".

Also, get rid of file path chooser label as it's not
needed, the text on before it explicitly instructs
to select the Android Source Package Directory.

Change-Id: Icc2319f177bd1839b22ab440242d536b50d6be34
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>