Revert "Fix documentation build if include paths contain generator expressions"
This reverts commit 6ec3c8f9c9fddfd0b44574a638be06ed592ef8d5.

The reason for introducing the function in Utils was already reverted in
91d0bf1993a1c72558b9f1e62e5c32ad88823967 .

The call of the function was broken, because of instead of passing a
list of includes as a single argument, it passed each include
as a separate argument (the subtle difference between
`${_include_dirs}` and `"${_include_dirs}"` for lists...).
Which subtly breaks the translation update.

Just go back to the original.

Change-Id: If745c434a9ff808e08703c3c1bf1393eb0183236
Reviewed-by: Christian Stenger <christian.stenger@qt.io>