Make qmlls semantic highlighting optional
Qmlls semantic highlighting in QtC occasionally flickers on editing the
document. Since this might be disturbing to some users, we should be
able to toggle its switch so that they can enable/disable it. By
default, it is disabled.

Add an option in the qml language server preferences to
switch highlighter module.

Add a check in highlighter() to test if there is actually semantic
highlighting request. Otherwise, it is invoked on the initialization of
the language client unconditionally which wouldn't respect the settings.

Task-number: QTCREATORBUG-31215
Change-Id: I1d7df44071053993839860087fb0ab6ede61e145
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>