Android: Fix crash on reloadPackages
The reloadPackages() deletes old packages, while the internal pointers
to these packages are still stored inside the AndroidSdkModel. Later,
sdkManagerCommand() executes in a blocking way a process with EventLoopMode::On,
causing the model view to repaint and take invalid data from the AndroidSdkModel.

The fix is to delete the packages after the nested event loop is
finished.

Fixes: QTCREATORBUG-31920
Change-Id: Id7e41f7d43ed299329974925c3dc246c4ffbf260
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>