QmlDesigner: Ignore component wit non existing files
Instead of throwing an exception we now silently ignoring non existing
components. But that has the disadvantage that we maybe later miss the
component if the the file is added later because the qmldir was not
changed.

If we want to handle that case too we have have to save more meta data.

Task-number: QDS-13954
Change-Id: Ia17207dadfd2c130d19d3316c2643a5e357c7074
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>