CMake: fix crash after changing the cmake build directory
The changes to the build directory line edit are applied automatically
on focus out events. If the new directory does not exist a message box
pops up asking the user whether to start a new configuration in the new
path, this message box spawns a new eventloop. If the focus out event is
triggered by opening a menu from the menubar a QAlphaWidget for the menu
fade in animation is used. This gets deleted in the messag box event
loop, but still gets used after the message box finishes.

Workaround this by disabling menu fading effects on Windows.

Fixes: QTCREATORBUG-31890
Task-number: QTBUG-130696
Change-Id: I98742136f3aa9fe5c42923478d9229a90c3eef05
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>