Utils: Fix insertion in EnvironmentModel
Previously when the user tried to add a new environment variable
to the list (e.g. in "Build Environment") it would add:

NEWVAR=VALUE
_=VALUE

Instead of the correct "NEWVAR=VALUE" only.

This was due to findInResultInsertPosition returning the position of
a match instead of the position one past the match.

Amends 727be63dacaff6a8327f2c74852a7d907f4ca7e3

Change-Id: I7ba5e059c532505a4fe24e77a0b8022737adaa20
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>