AutoTest: Fix execution of tests with multiple test classes
In case of multiple test classes inside a single application
we already limited the execution to run all as it is
impossible to address the respective test separately.
But when iterating over all test cases the fact of multiple
test classes was ignored completely which lead to running
the same tests for the number of used classes.
Ensure a single target gets executed only once.

Fixes: QTCREATORBUG-31935
Change-Id: I386560ff10ead2884c170670fc41df07234202f1
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>