Find: Fix reading settings
It was not possible to go back to the default filter after other filters
got added to the list. Saving and restoring the current filter was not
symmetric for the default case: When writing, the current value was not
written if it was default, when reading the current value, it was set to
the first value in the list if not set in the settings, instead of the
default.

Fixes: QTCREATORBUG-31676
Change-Id: I5cde924f56744f9142b5cc52b51a9f08d06b3d64
Reviewed-by: David Schulz <david.schulz@qt.io>