Lua: Fix crash from settings created in coroutine
When calling dynamic_set, or create({...}) from a coroutine
the objects would live on the coroutine thread. Once that
coroutine ended, the objects would be garbage collected.

When a script later tried to access the stored objects a crash
would occur, since the objects were no longer valid.

To keep this from happening we make sure to move the objects
onto the main thread which lives as long as the plugin does.

Change-Id: I5340f16429a51b02a5cef0d93a76c473173faf54
Reviewed-by: hjk <hjk@qt.io>