QmlPreview: Fix resetting tasktree from its handler
It may happen that we are resetting the running task tree
directly from one of its handler, which may lead to crash.
The code path may look like:
1. onParserDone [qmlpreviewplugin.cpp : 508]
   task tree is still running
2. QmlPreviewPluginPrivate::triggerPreview
3. QmlPreviewPlugin::previewCurrentFile
4. QmlPreviewPlugin::setPreviewedFile
5. QmlPreviewPlugin::previewedFileChanged
6. QmlPreviewPluginPrivate::checkFile
7. QmlPreviewPluginPrivate::checkDocument
8. and now we are resetting the still running task tree.

Fix it by providing TaskTreeRunner's done handler,
which is safe to be used in such a scenario, as before
the handler is invoked, the finished task tree is already
scheduled for deletion by a call to deleteLater().

Add a test for TaskTreeRunner confirming such a scenario
is valid.

Amends bbed1896561b2f62449cd4f87ef95459c93d2fa1

Change-Id: I8671ffd4e82c505c6ed98687948db1273114e65a
Reviewed-by: Eike Ziller <eike.ziller@qt.io>