Find: Fix tab order
Removing a widget and adding a different widget messes up the tab order.
Reinstate a nice tab order explicitly. We have to reset the tab order on
all the involved widgets, and need to skip widgets with NoFocus focus
policy (even though they appear in the original tab order) because
setTabOrder is a no-op if any of the widgets do have NoFocus.

Fixes: QTCREATORBUG-31771
Change-Id: I1db45f81ed37b059d46641d59afd43fc785830b7
Reviewed-by: Christian Stenger <christian.stenger@qt.io>