Build: Fix zipping of dev package on Linux
Official 7zip normally checks for the existence of symlink targets, and
fails if they do not. That can be avoided with a command line flag. This
is important for the dev packages on Linux, where the libFoo.so is a
symlink to the actual library, and the symlink is part of the dev
package, but the actual library part of the non-dev package.

Change-Id: Ibd5855c4e106c96058a693d87e84d0f2d8f93365
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Patrik Teivonen <patrik.teivonen@qt.io>