Debugger: pmr containers
Added support for visualization of for `std::pmr` containers
(and more broadly containers with custom non-empty allocators).
The gotcha with pmr allocators in particular is that
unlike most allocators these are non-empty classes:
they contain a pointer to a memory resource.

Tested locally with CDB and GDB (MinGW).

Task number: QTCREATORBUG-30224

Change-Id: If0ceb8a46905151a98f27d3e3ff11386406a9737
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>