MesonProjectManager: Add MIME entry for meson.{build,options}
The Apache Tika MIME database in use since commit
36dca9eb8c0018b6a3cb802e35a9ce3f36786ef8
currently doesn't contain an entry to be able to identify
that "meson.build" is a meson build description.

Add a corresponding MIME entry in order to detect a MIME type
of "text/x-meson" to match what the MesonProjectManager declares
support for (s.
MesonBuildConfigurationFactory::MesonBuildConfigurationFactory).

In the same way, add "meson.options" and "meson_options.txt" (which
is used with meson versions < 1.1, see [1]), which can also
be used to load meson projects.

With this in place, selecting a "build.meson" or "meson.options" file
to load a meson project works on Linux when the experimental
MesonProjectManager plugin is enabled, instead of showing an error
message:

> Failed opening project ".../path/to/meson.build":
> No plugin can open project type "text/plain".

[1] https://mesonbuild.com/Build-options.html

Change-Id: I328c9d23e84a4b07e7c7cd0b296f72af441ce54c
Reviewed-by: Alexis Jeandet <alexis.jeandet@member.fsf.org>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>