Git: Allow blame for real single line selection
Before, the selection needed to span two lines, at least
the cursor had to be in the next line. This was to not
trigger a single line blame when a word was selection.

But now allow a selection from line start to line end also.

Amends 7a5b45bdc22ff10ba52fc1b31f2c4bb7819f19a0

Change-Id: I629a551a98156d372ac7265faad907a3185ab258
Reviewed-by: Orgad Shaneh <orgads@gmail.com>