TaskTree: Fix withCancel test flakiness (2nd try)
Ensure the created task lasts long enough so that the 2 queued
signals with 0 timeout are delivered before the task finishes.

In case of a very slow machine, or when running valgrind memory analyzer,
it may happen that the time elapsed since queueing a call, until
the call is handled, may be longer than 1ms. In this case, the
task's done handler is invoked before the cancel signal is delivered.

To ensure that it doesn't happen, make the task last for very long time.

Amends 30d3f8ceb810705671b6b132f5a5fe5b1197d582

Change-Id: Icdeb48df6ed385eab86e0447baca128ccf50de37
Reviewed-by: hjk <hjk@qt.io>