CppEditor: Fix "move definition" edge case
Consider the following class:

class C {
  C();
  C(const C&) {}
};

Note that the default constructor is unimplemented.
If you now tried to move the second constructor's definition outside the
class (or indeed any other member function following the two
constructors), then the definition would end up inside the class again,
but wrongly qualified, thus creating invalid code. Here's what happened:
  1. To determine the best insertion point for the moved definition,
     we try to locate definitions of the functions around the one
     to be moved.
  2. For this purpose, SymbolFinder::findMatchingDefinition() is called
     for the default constructor.
  3. There is no such definition, but because the "strict" parameter
     was wrongly set to false (indirectly via a default argument), we
     find the other constructor's definition and believe this to be a
     suitable location for the moved definition.
We fix this by changing the value of the "strict" parameter to true.
We also remove that argument's default value, because it is harmful:
Every caller should make a conscious decision about whether the look-up
should be strict or not.

Change-Id: Ieb635d817ea0d0b5eabf7a9cbc0d90b4cf99626d
Reviewed-by: Christian Stenger <christian.stenger@qt.io>