AutoTest: Avoid freeze on changing sort order
Do not use the link data for handling the sort order
as e.g. Qt Tests have a specialized functionality which
makes the link creation quite expensive and should
always only be done for a single item.
Use the information we have instead of looking up some
more correct locations.
Minimal behavior change as this can change the sort
order slightly as it is now ordering by the location of
the declaration instead of the definition.

Fixes: QTCREATORBUG-31994
Change-Id: I5f5e956ef02477845b8bea7c8d2897b5ebd0f278
Reviewed-by: David Schulz <david.schulz@qt.io>