RemoteLinux: Introduce an indirection in the device
Plan is to handle the "disconnected" state as a separate type of access
which wouldn't have the shell thread. Changing states would happen by
destructing an existing access object (consisting of a file access
object and surrounding helpers/mutex/...) and creating a new one, so it
makes sense to have any state-internal data/objects part of the access
object, so they can be constructed/destructed as a unit.

Change-Id: I3fdb40b560e5902a40c96703afb8678c2f60c80d
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>