VCS: Allow disabling reflection of file status
Displaying the file vcs status permanently may become quite
expensive.
Explicitly let the user decide whether to enable this feature
or not, but disable it by default for now as there should
be several ways to improve performance and limit running
the vcs client to times when this is really necessary.

Task-number: QTCREATORBUG-32002
Change-Id: I00d206d52b74b5daa5b5d0b0022b32f11d51faf5
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>