Plugins: Change a few usages of PluginSpec::name to ::id
First, PluginSpec had only field "name". QtC 15 added "id", which became
(a lowercase) copy of "name", and which is the future-proof field to use
when identifying a plugin.

This change turns usages of PluginSpec::name, which are not exclusively
used for user display into PluginSpec::id. In case of id string
comparisons, the plugin id string turns to lowercase.

Change-Id: Ibae3b06a932158cd6946e1b29192b1db0dd78a40
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>