LanguageClient: fix hasItemsToPropose check in completion model
Prefilter the candidates before checking whether the item is a perfect
match allows to ignore items that are part of the inserted code to
complete. In the example of QTCREATORBUG-32013 tex matched against x and
returned a perfect match in that case. If the model reports a perfect
match the completion widget is not shown by default but needs to be
triggered explicitly.

Fixes: QTCREATORBUG-32013
Change-Id: Icd1da6163f0bee2ff09a37456dd51b350fd4fe10
Reviewed-by: Christian Stenger <christian.stenger@qt.io>