KArchive: Add KArchive library and its dependencies
We want to include a reliable way to extract archives.
KArchive is a library that provides a simple API for reading and writing
archives. It will allow us to replace the current implementation of
archive extraction which relies on the necessary tools
(7zip, unzip, tar etc.) to be available on the users system.

Since karchive is built on top of QFile if will also allow us to
use it for reading and writing archives on remote devices.

Change-Id: I390e0e786f2dd20ae85d16dab34f43510cdf82a6
Reviewed-by: Eike Ziller <eike.ziller@qt.io>