fix: [session] The `default` session can be deleted and renamed
When opening the session manager for the first time, the default session can be deleted and renamed.

The default session is selected during construction, and the active session is selected in the showEvent.
If the active session is the default session, the selectionChanged signal won't be triggered, which leads to the default session being able to be deleted and renamed.

Change-Id: I2186d02e93bcd3243f3ccf76645c19224f3d32db
Reviewed-by: Eike Ziller <eike.ziller@qt.io>