Tooltips: More sensible position if it would go out of screen
The code that ensures that a tool tip is visible on screen as
much as possible, first tried "flipping" the tooltip horizontally and
vertically, so a tooltip would be visible to the left or to the top of
the cursor instead of to the right and bottom.

While that makes sense for the vertical position, because the tooltip
would get under the cursor if it was just moved inside the screen, this
is a really weird behavior horizontally. The tool tip is already
vertically separated from the cursor (it is further below), so we can
just move the tooltip as little as possible horizontally to make it
fully visible on the screen.

Change-Id: I25e70c9a57602992c1a2fa1015cc8f23d9cb6f78
Reviewed-by: David Schulz <david.schulz@qt.io>