Tooltips: Fix widget tool tip positioning on macOS
There is a weird behavior of Qt on macOS, that widget tooltips are moved
up by 20 pixels when calling show().
This is pretty annoying, because the tooltip gets under the mouse
pointer and interrupts user interaction.

This only happens if all of

- the widget has multiple subwidgets
- the layout has size constraint QLayout::SetFixedSize
- adjustSize is called

are true.

Remove the explicit adjustSize, that doesn't seem to make any
difference, except for avoiding the issue with show().

Task-number: QTBUG-131479
Change-Id: I2285bf77b0980940b48ac22f931a45b6a73ad06e
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>