VCS: Fix issues with clicking on links in output view
Amends cb6a2da37223fa34f4e173e2ed64118f735d6c92

that started clearing the static repository information from the output
view. The issue with that is that it clears that information even for
asynchronous vcs commands that do not have anything to do with the
output view at all. The code before was mostly wrong already, but the
information was never cleared it mostly worked out.

It would be good to get rid of the static information about the
repository from the output view, and instead provid the repository
information for each output that is added. But this patch is a band aid
for Qt Creator 15:

- set the repository information again before appending messages from a
vcs command to the output view
- set the repository information also for runBlocking, which also broke
in the above commit

Task-number: QTCREATORBUG-31894
Change-Id: I0823b53e0fbecc14961be905f5279eb5a461ab93
Reviewed-by: André Hartmann <aha_1980@gmx.de>