Terminal: Fix "flicker" while searching
Previously the search results would be removed every
time the terminal updates before the new hits were
resolved.

This patch fixed that by keeping the old hits until the
new hits are found.

To not keep old hits around when the terminal is cleared
a signal is introduced that the SearchableTerminal can connect
to and clear its Hits immediately.

Fixes: QTCREATORBUG-32066
Change-Id: If0bcf133098f06df53b52aae9b1036f0bc58870d
Reviewed-by: Cristian Adam <cristian.adam@qt.io>