Editor: use detected indentation by default
This removes the undescribed Mixed tab settings option. It was unclear
and undocumented what this option actually means.

In theory this option tries to detect whether tabs or spaces are used
for indentation around a specific position in the document. This now
conflicts with the automatic auto detection which scans the complete
document. That auto detection not just detects whether tabs are used but
also the indent depth. Having a source document with a different indent
depth is arguably a more common use case than a source file with mixed
tabs and spaces for indentation. So in order to not confuse the user
with to many unclear magic options that might interfere with each other
the mixed tab settings option was removed in this change.

Fixes: QTCREATORBUG-11575
Fixes: QTCREATORBUG-11675
Fixes: QTCREATORBUG-19576
Fixes: QTCREATORBUG-25628
Change-Id: Ib95662ade38d0384d503e9a7b99f54ea4b416f68
Reviewed-by: Christian Stenger <christian.stenger@qt.io>