VCS: Detect VCS directories even for disabled VCS plugins
Most of the version control support plugins detect that a directory/file
is under version control by looking for certain files in the directory
structure. This search doesn't really need anything specific from the
plugin, except for a list of file names.

Allow version control plugins to specify the list of files to look for in
their plugin meta data as "VcsDetectionFiles". When it is checked if a
directory is under version control, and none is found from the enabled
plugins, use the meta data to find out if any installed but disabled
plugin feels responsible for that directory.

Show a notification if a plugin is found that handles such a detected
version control system, with the option to enable the plugin and restart
QtC if necessary.

Since this adds discoverability of the version control support even when
plugins are disabled, disable the VCS plugins that can use this
mechanism (except for Git).

Change-Id: Ib507572c0065dd889a2f9b780c794f9cd985e265
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>