LSP: recreate clients enabled on a project
Add a method on BaseSettings to check whether it is enabled on a certain
project via global or project settings, and use it when creating a
client to create globally disabled but project-specifically enabled
clients.

Amends 18cd9359b9378dca470a04c782cca4d35ac8636c that introduced the
enabling/disabling of settings project-wide.

Change-Id: I1e6f2212ef6da965d01bd09348238c98a3a768ba
Reviewed-by: David Schulz <david.schulz@qt.io>