qml2puppet: update pixmap when clipping changes
This is a fix for a corner case in the form editor when an element is drawn out of it's bounding rectangle.
Pixmap update was triggered only on visibility changes (not on clipping).
Due to lack of information about pixmap size changes, item was stucking at the bounding rectangle
when it became visible and did not react to clipping changes in the editor.

Task-number: QDS-14044
Change-Id: If074ca6c1b11216c0be12a4fb44e4e79688d6b5c
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>