Utils: Use QImage instead of QPixmap for current frame in QtcMovie
The current frame was internally stored as QPixmap, and in all cases
unconditionally created from a QImage. This caused a QImage -> QPixmap
-> QImage roundtrip for currentImage(). We now store the frame as
QImage, effectively removing the round trip when using currentImage()
and delaying the conversion when using currentPixmap().

A usage pattern of repeated calls to currentPixmap() which naively would
result in multiple calls of QPixmap::fromImage() on the same image is
counter-acted by caching and later re-using the first call's result.

Change-Id: I1a76aee523ed289a266bd15511ad9f07803abdcf
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>