QbsProjectManager: Force full resolving on "Reparse Qbs"
This option is intended for "trying again" if something failed in a way
that changes to the problematic file won't get picked up because it was
so broken that it did not end up in the list of build system files to
watch. However, if that is the case, then qbs' own change tracking
mechanism will also not detect a change, so the whole procedure was
useless.
We fix this by explicitly instrucing qbs to do a full (re-)resolve for
that particular action.

Change-Id: I8d686e5e5a4312c3c8f23498edd4af871090e96e
Reviewed-by: Christian Stenger <christian.stenger@qt.io>