ProjectExplorer: Create a thin DeviceRef wrapper class
... around a weak pointer to IDevice to pass around device identity.

The idea is to ramp down the direct use, especially storage, of
IDevice::{Const,}Ptr outside the DeviceManager and use DeviceRef
instead, in order to avoid the problem of device shared pointers
not dying when copies are held in dormant dialogs or similar.

The separate class effectively acts as a weak pointer with somewhat
nicer syntax.

Change-Id: If735ed8af9589137640b73b023d04c063ea876be
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>