Debugger: Add an option to control heap debugging
Added a checkbox (temporarily) in CDB settings page that toggles
the heap debugging on Windows.
Although it is "technically" a CDB option, it woks with LLDB on Windows
as well.
The default behavior is to keep the heap debugging off.
The checkbox can still be overridden by the environment variable.

Fixes: QTCREATORBUG-32102
Change-Id: I580cd0d66da38776f3a51632ede6b5f60d4d4cfd
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>