EffectComposer: Prefer fragment-shader tab on code editor
The fragment shader is opened by default when it's not empty.
Also, when both shaders are empty, the fragment shader is prior to being
selected by default.
If a node doesn't have a fragment shader, but has a vertex shader, the
vertex tab is preferred to be selected on opening.

Fixes: QDS-14319
Change-Id: I2ec7e36fc9c739e1849a412d22751c2d5b4867e4
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>