DeviceManager: Create initial device id
Before this patch Device Manager was depending on the device id sent by
the device. But it was already needed for internal process
before the device actually connects and sends it. This patch
generates a device id when the device is added to the system and
distinguishes it the device's self id.

Task-number: QDS-14269
Change-Id: Ie6417975064423ddac8a255a0ee720ab2637cb11
Reviewed-by: Henning Gründl <henning.gruendl@qt.io>