Utils: Use Qt::ColorScheme-based API
The OS specific handling of light/dark system theming in Qt Creator was
implemented by passing around bool values. Qt 6.5 comes with a new enum
Qt::ColorScheme and respecive themeing API.

This change replaces some of the legacy code with the new Qt API.

"System Information" displays the color schemes of the current theme and
the one of the system.

Change-Id: Id39af399dd54ea239bd750c45ea74a48be5efbd2
Reviewed-by: Cristian Adam <cristian.adam@qt.io>