Fix compile with Qt 6.9
The QString::arg overload that accepts a base is explicitly disabled for
everything that can be cast to a QAnyStringView, so for example
T==char32_t or T==char16_t will is not usable in that overload. And the
overload that accepts those types does not allow to set the base. Fix
this by using T==int32_t or T==int16_t.

Change-Id: Ib23a2ced7e03dab6de50458d36bc27689c6f8da9
Reviewed-by: hjk <hjk@qt.io>