Fix installing just a library in plugin install wizard
It is possible to just select a plugin library in the plugin install
wizard, and that worked by just copying the library over.

Amends be2e5262c73ef747511a10e04a36338d3d49412e
that introduced a crash for that case, by unconditionally using the
PluginSpec instance that was only created for the "archive" case.

Instead of skipping the check archive step, also try to read a
PluginSpec for pure library files. That has the advantage that we do some
basic checks for them too, and that we have the PluginSpec instance
available for the later code.

Change-Id: I23286d81de54c08d4735c9bcdb2368c8b6097408
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>