RemoteLinux: Drop GenericDeployStep's not-same-device restriction
This is getting into the way for remote builds as it disabled
the GenericCopy version which is in principle always usable.

Make this even the preferred version for the build == run case,
as this will end up in the fast path of FilePath::copyFile.

As a collateral improvement this also removes the error message
that was referring to rsync only.

Change-Id: Ia27e3733b99f0de0161050c0355f135026b899f5
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>