QmlProfiler: Fix crash at startup if file creation fails
If opening the temporary file for QmlProfilerEventStorage fails in it's
destructor, it was crashing, because the error handler tried to emit a
signal on an object that wasn't even half constructed yet.

Can be tested by forcing the return value of TraceStashFile::open to
`false`.

Delay the construction of the event storage a bit and do not set the
error handler in its constructor, since nobody could connect to the
signal that was sent yet anyway. Also add some checks before calling the
error handler.

Change-Id: I0fc1207aac447090e12f6b2342e156312e0d1d1b
Reviewed-by: hjk <hjk@qt.io>