languageclientformatter: prepare for supporting full document formatting
languageclientformatter used to support only range based formatting
prior to this commit. In qmlls, range based formatting is too limited
and we would like to use the power of full document formatting.

Preparing full document formatting request is quite similar to range
formatting request. Isolate the common part in an helper. Wrap format
requester in an interface so that we decide what kind of request we want
at runtime.

Also modify reformat() of qmljseditor so that it first try to format the
file using the language server.

Task-number: QTCREATORBUG-26602
Change-Id: I4f722eb04e68ab83848548ecadfbf1dcde446ba5
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>