ProjectExplorer: (Mostly) implement IDevice::filePath on top of rootPath
Separate implementation led to mixed results depending on the exact
path of creation, leading to confusion in the debugger when paths
of files reported by the debugger were matched with already
opened files in the EditorManager.

DesktopDevice is the odd-one out due to the fixed C:\ rootPath() on Windows
and filePath() not using (and possibly wanting) this.

Change-Id: I02858747524107124189498c777cc79b546ecce9
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>