Android: Don't show version of selected NDK in status summary
When opening the options page, the first entry in the NDK list widget got automatically selected. The version of that NDK entry was displayed in the status summary, and the status summary got updated when the
selection changed.

That gave the impression that the selection of an NDK entry has an
influence on the configuration of kits. That is not the case, though.
The NDK list widget is there to explore and manage available NDKs, and
the kit generation mechanism has its own criteria to determine which of
those NDKs to use for which kit.

This change removes the automatic selection of the first NDK entry, as-
well as the displaying of the NDK version in the status summary. The
translations of the status summary string are also adapted.

Fixes: QTCREATORBUG-31900
Change-Id: Ia5e8d8401bf4e3e76093dc7341198cc8c8678214
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>