Qbs: Return FilePath from QbsSettings::qbsSettingsBaseDir
Allows using .nativePath() with the right context when needed.

Currently this does not change anything, and most users are ok
with using .path() when .nativePath() would be "correct" anyway.

This also needs some further thinking on where to pick up the
settings when building remotely, but at least one explicit use
of FilePath::toString() and one of QDir::toNativeSeparator() are
removed without ill-affecting the current local use case.

Change-Id: I23566d0909181b44737c381624ec4466854d4ce4
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>