QtcSeparateDebugInfo: Adopt dsym location change from Qt
Fix install location of dSYM bundle for Qt apps
a60d42d873e67281e5e30bcaa3daa2f8edffe7eb

This actually doesn't change anything in the installation of Qt Creator,
only the build-time location. We do not use the standard install target
for app bundles of CMake, which just copies the app bundle, which is
ugly if the app bundle contains the .dSYM directories. This change keeps
the .dSYM out of the app bundle in the build directory, which is still a
sensible thing to do.

Change-Id: I42cc8875e1e123a79a810d2b2e5dde8c2e389c53
Reviewed-by: Cristian Adam <cristian.adam@qt.io>